+++ zcode/plot/plot_const.py
@@ -1,9 +1,0 @@
-"""
-"""
-
-LW_CONF = 1.0
-LW_OUTLINE = 0.6
-
-COL_CORR = 'royalblue'
-
-_PAD = 0.01